Support dynamic partition sizing in kernel/rootfs creation.
This patch enables openwrt to calculate in flash partition sizes for the
zipit. Partition sizes are calculated and a new dtb created for the
kernel. It allows for easier rootfs customization; adding or removing
packages to squashfs without a constrained partition size.

A new mtdblock named "firmware" is added to the kernel. It begins at the
same address as kernel and ends at the end of the internal flash. This
allows flashing the complete kernel/squashfs/jffs2 as a single "firmware"
file.

The last stage of image creation checks the firmware image size and if
the image is too large for the internal Zipit flash it will exit the
build proccess with an error.

Signed-off-by: Joe Honold <mozzwald@gmail.com>